Testing Framework now supports faking of object tracker returns/delays
- Supports faking of object tracker returns/delays
- Controller test suite now supports control and target clients
